requirements.txt
selenium>=3.14.0
names>=0.3.0
tbselenium>=0.3.3
psycopg2-binary>=2.7.5
